Check for duplicates when adding a declaration to a scope.
Special care is taken to allow function overloads, but not overloading functions with other declarations.

Added add_declaration helper function to parser. It may make sense to use this in other locations, such as in the polymorphic function arguments.

Also, should we allow overriding declarations in function argument scope with declarations in the function body?