Add support for custom iterator declarations in for loops.
This also fixes a minor bug in the declaration of iterator when using indexed expressions:

    decl->identifier->enclosing_scope = &_for->body;

Iterator index not supported yet. Not sure which one of these we should use:

    for it_index,it in array
    for it,it_index in array
    for (it,it_index) in array
    for (it_index,it) in array

I like the index first because it maps better to key,value pairs. Should we parenthesize the expression like in swift? Does that mean that it would be a tuple? Should we support both?